Add files via upload
Dataset was originated from staff at OCLC Research and shared with Eira Tansey in JSON. JSON was uploaded to Refine and edited to remove rows that served as headers for U.S. states (contained no repository data), as well as rows related to non-American repositories (primarily in Canada and Europe). Data was then downloaded from Refine in CSV format.